[refactor] [design] 커스텀 훅 분리, 포트폴리오 카드 디자인 수정
