+++ src/components/PortfolioCard.module.css
@@ -9,15 +9,15 @@
 
 .container {
   display: flex;
-  padding: var(--sp-2) var(--sp-3);
+  padding: 8px 8px 8px 12px;
   flex-direction: column;
-  align-items: flex-start;
   gap: 16px;
   flex-shrink: 0;
   border-radius: 6px;
   border: 1px solid var(--gray-50);
   background: var(--white);
   position: relative;
+  min-height: 240px;
 
   h3 {
     font-size: 20px;
@@ -26,9 +26,19 @@
   }
 
   .title {
-    display: flex;
-    justify-content: space-between;
-    flex-direction: row;
+    overflow: hidden;
+    text-overflow: ellipsis;
+    display: -webkit-box;
+    -webkit-line-clamp: 2;
+    -webkit-box-orient: vertical;
+  }
+
+  .content {
+    overflow: hidden;
+    text-overflow: ellipsis;
+    display: -webkit-box;
+    -webkit-line-clamp: 1;
+    -webkit-box-orient: vertical;
   }
 
   .interest {
@@ -37,15 +47,25 @@
     padding: 4px 12px;
     border-radius: 6px;
     font-size: 14px;
+    max-width: 130px;
+    text-align: center;
   }
 
   .bookmark {
-    position: absolute;
-    top: 12px;
-    right: 12px;
     border: none;
     background-color: transparent;
     cursor: pointer;
+    position: absolute;
+    top: 4px;
+    right: 0px;
+  }
+
+  .career {
+    color: var(--gray);
+  }
+
+  .career::after {
+    content: ' | ';
   }
 
   .meta {
@@ -53,7 +73,19 @@
     flex-direction: row;
     justify-content: flex-end;
     align-items: center;
-    gap: 8px;
+    gap: 12px;
     width: 100%;
+    position: absolute;
+    bottom: 8px;
+    right: 10px;
+
+    .like,
+    .view {
+      display: flex;
+      flex-direction: row;
+      justify-content: center;
+      align-items: center;
+      gap: 8px;
+    }
   }
 }
